Move json to string to common pipe
